Time: 0 ms (100%), Space: 14 MB (31.05%) - LeetHub

+++ 0260-single-number-iii/0260-single-number-iii.cpp
@@ -7,7 +7,7 @@ public:
         }
 
         // Get rightmost set bit (the bit where the two unique numbers differ)
- int differenceCreator = xORR & -(unsigned int)xORR;
+ int differenceCreator = xORR & -(long long)xORR;
 
         int bucket1 = 0, bucket2 = 0;
         for (int num : nums) {
